docker-image/locust-tasks/requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
ConfigArgParse==1.5.2
Flask==2.0.1
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
greenlet==1.1.1
gevent==21.8.0
geventhttpclient==1.4.4
idna==3.2
importlib-metadata==4.6.3
itsdangerous==2.0.1
Jinja2==3.0.1
locust==2.1.0
MarkupSafe==2.0.1
msgpack==1.0.2
psutil==5.8.0
pyzmq==22.2.1
requests==2.26.0
roundrobin==0.0.2
six==1.16.0
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
zipp==3.5.0
zope.event==4.5.0
zope.interface==5.4.0
